groomer cards now rendering with search function
