+++ src/components/pages/GroomerCards/GroomerCards.js
@@ -3,20 +3,10 @@ import { getGroomers } from '../../../api';
 import { Card } from 'antd';
 import './groomerCard-styles.scss';
 
-function GroomerCards() {
-  const [data, setData] = useState([]);
-  const fetchdata = async () => {
-    const groomers = await getGroomers();
-    console.log(groomers);
-    setData(groomers.data);
-  };
-  useEffect(() => {
-    fetchdata();
-  }, []);
-
+function GroomerCards({ groomers }) {
   return (
     <div className="groomerWrapper">
-      {data.map(e => (
+      {groomers.map(e => (
         <Card className="groomerCard" bordered={false}>
           <h1>{e.business_name}</h1>
           <img className="groomer-img" src={e.profile_picture} alt="groomers" />
